Add scrapper fallback to old mappings
Usually, after an update to UONET+ it will take anywhere from several minutes to hours to update the mappings for the Scrapper. However, sometimes, the mappings remain the same. The Scrapper will now attempt to use the latest avaliable mappings, even if there are no mappings for the current UONET version.